requirements.txt
# Google AI
google-generativeai>=0.8.0

# FastAPI for backend
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Data validation
pydantic>=2.9.0
pydantic-settings>=2.6.0

# Utilities
python-dotenv>=1.0.0
requests>=2.32.0

# Google Cloud (optional for production)
google-auth>=2.23.0
